Endangered Sealife Description
This is the whole game. In each section there are the different assets
that are being used. Such assets are the materials, scripts, textures,
prefabs, sprites, animation, 3D objects, environment, different scenes
and sound. Once the game is completed the whole folder will be updated
so that one could proceed with a pull request.