sound, "view your character button" and jump button
1. The jump action was adjusted so that the user can use the mouse to
jump, also the code had to be adjusted because when the user would
select the button the character would keep on rising up in the air when
i needed it to just jump a certain height on one click. Also more sound
was added to the assets such as main menu music. Lastly a page was set
so that the users can view the character they'd be using.