Code for the Wolf controller(Jump Action )
the code for the wolf controller was modified so that the character will
only jump if the wolf is first in contact with the ground. If the Wolf
is not in contact with the ground he will not be able to jump. This was
an issue because the user was able to double jump several times if
he/she pressed the jump button many times in a row)